+++ client/src/api/verification.js
@@ -1,11 +1,11 @@
 import { post, get, patch, remove } from './base'
 
 export async function verifyUser (userId) {
-  return await patch(`/user/${userId}/verification`)
+  return patch(`/user/${userId}/verification`)
 }
 
 export async function deverifyUser (userId) {
-  return await remove(`/user/${userId}/verification`)
+  return remove(`/user/${userId}/verification`)
 }
 
 export async function getVerificationHistory (userId) {