+++ CHANGELOG.md
@@ -56,7 +56,7 @@
 - put Date and time in two lines at User Dashboard Card "Nächste Abholungen". #1448 !2588 @Niklas003
 
 - Replaced the history dialogs on profile pages with a vue component !2587 @alex.simm
-- Moved verification to memberlist !2646 @chriswalg
+- Moved verification to memberlist !2646 !2698 @chriswalg
 - The broadcast banner fetches its content from the API #1284 !2661 @alex.simm
 - added poppins font in Styleguide !2564 @Niklas003
 - Replace the region tree with a new vue component !2483 @alex.simm